Blacklisted schema.rb as it is frequently causing merge conflicts.
This file changes with every migration run as such it is different
across space, time, and even branches. Although it is recommended
to have this file committed when it changes, it is causing more
problems in development than it is actually solving.

NOTE: Will be a good idea if it is whitelisted when moving to
production.