Fixed leak of logged in user and location across threads.
The logged in user and the location the user logged in from are stored
in the global User.current and Location.current variables for every
request. It's an easy way to pass those details between controllers and other
components of the application (the application ought to get refactored
to have these injected into those other components). However, this was
implemented in a naive way that allowed different threads to share the
variables. Wasn't much of a problem in PoC sites that run the application
using the non-enterprise version of passenger which is not multi-threaded
(only multiprocessed) but it could potentially cause problems in other
deployments using multithreaded servers like Puma (eg all eMastercard
deployments use Puma).